Refactor export code
